refactor: extract services queries in dedicated lib
