+++ libs/domains/application/src/lib/slices/applications.slice.ts
@@ -15,10 +15,6 @@ import {
   ApplicationMetricsApi,
   ApplicationRequest,
   ApplicationsApi,
-  CloneApplicationRequest,
-  CloneContainerRequest,
-  CloneDatabaseRequest,
-  CloneJobRequest,
   Commit,
   ContainerAdvancedSettings,
   ContainerConfigurationApi,
@@ -27,7 +23,6 @@ import {
   ContainerMetricsApi,
   ContainerRequest,
   ContainersApi,
-  DatabasesApi,
   DeploymentHistoryApplication,
   Instance,
   JobAdvancedSettings,
@@ -39,7 +34,6 @@ import {
   Link,
   Status,
 } from 'qovery-typescript-axios'
-import { useMutation } from 'react-query'
 import { ServiceTypeEnum, isApplication, isContainer, isJob } from '@qovery/shared/enums'
 import {
   AdvancedSettings,
@@ -80,60 +74,10 @@ const containerMetricsApi = new ContainerMetricsApi()
 const containerDeploymentsApi = new ContainerDeploymentHistoryApi()
 const containerConfigurationApi = new ContainerConfigurationApi()
 
-const databasesApi = new DatabasesApi()
-
 const jobsApi = new JobsApi()
 const jobMainCallsApi = new JobMainCallsApi()
 const jobDeploymentsApi = new JobDeploymentHistoryApi()
 const jobConfigurationApi = new JobConfigurationApi()
-
-type UseCloneServiceProps =
-  | {
-      serviceId: string
-      serviceType: ServiceTypeEnum.APPLICATION
-      cloneRequest: CloneApplicationRequest
-    }
-  | {
-      serviceId: string
-      serviceType: ServiceTypeEnum.CONTAINER
-      cloneRequest: CloneContainerRequest
-    }
-  | {
-      serviceId: string
-      serviceType: ServiceTypeEnum.DATABASE
-      cloneRequest: CloneDatabaseRequest
-    }
-  | {
-      serviceId: string
-      serviceType: ServiceTypeEnum.JOB | ServiceTypeEnum.LIFECYCLE_JOB | ServiceTypeEnum.CRON_JOB
-      cloneRequest: CloneJobRequest
-    }
-
-export const useCloneService = () => {
-  return useMutation(
-    async ({ serviceId, serviceType, cloneRequest }: UseCloneServiceProps) => {
-      const mapper = {
-        [ServiceTypeEnum.APPLICATION]: applicationsApi.cloneApplication.bind(applicationsApi),
-        [ServiceTypeEnum.CONTAINER]: containersApi.cloneContainer.bind(containersApi),
-        [ServiceTypeEnum.DATABASE]: databasesApi.cloneDatabase.bind(databasesApi),
-        [ServiceTypeEnum.JOB]: jobsApi.cloneJob.bind(jobsApi),
-        [ServiceTypeEnum.CRON_JOB]: jobsApi.cloneJob.bind(jobsApi),
-        [ServiceTypeEnum.LIFECYCLE_JOB]: jobsApi.cloneJob.bind(jobsApi),
-      } as const
-      const mutation = mapper[serviceType]
-      const response = await mutation(serviceId, cloneRequest)
-      return response.data
-    },
-    {
-      onSuccess() {
-        toast(ToastEnum.SUCCESS, 'Your service is cloned')
-      },
-      onError: (err) => {
-        toastError(err as Error)
-      },
-    }
-  )
-}
 
 export const fetchApplications = createAsyncThunk<
   ApplicationEntity[],
