+++ libs/domains/services/README.md
@@ -1,0 +1,7 @@
+# domains-services
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test domains-services` to execute the unit tests via [Jest](https://jestjs.io).
