+++ libs/domains/services/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/slices/services.queries'
