+++ libs/shared/console-shared/src/lib/clone-service-modal/feature/clone-service-modal-feature.tsx
@@ -1,9 +1,10 @@
 import { FormProvider, useForm } from 'react-hook-form'
 import { useDispatch } from 'react-redux'
 import { useNavigate } from 'react-router-dom'
-import { fetchApplications, useCloneService } from '@qovery/domains/application'
+import { fetchApplications } from '@qovery/domains/application'
 import { fetchDatabases } from '@qovery/domains/database'
 import { useFetchEnvironments } from '@qovery/domains/environment'
+import { useCloneService } from '@qovery/domains/services'
 import { ServiceTypeEnum, getServiceType } from '@qovery/shared/enums'
 import { ApplicationEntity, DatabaseEntity } from '@qovery/shared/interfaces'
 import { APPLICATION_GENERAL_URL, APPLICATION_URL, DATABASE_GENERAL_URL, DATABASE_URL } from '@qovery/shared/routes'
