+++ tsconfig.base.json
@@ -24,6 +24,7 @@
       "@qovery/domains/event": ["libs/domains/event/src/index.ts"],
       "@qovery/domains/organization": ["libs/domains/organization/src/index.ts"],
       "@qovery/domains/projects": ["libs/domains/projects/src/index.ts"],
+      "@qovery/domains/services": ["libs/domains/services/src/index.ts"],
       "@qovery/domains/user": ["libs/domains/user/src/index.ts"],
       "@qovery/pages/application": ["libs/pages/application/src/index.ts"],
       "@qovery/pages/cluster": ["libs/pages/cluster/src/index.ts"],
